Switch from PyXML to xml.etree
PyXML is an old library that's hard to install in modern Python
distributions. This commit modifies the XML parsing to instead leverage
the standard library xml.etree module [1]. The module has been around
since Python 2.5 which means that the updated code should run on the
vast majority of modern systems.

[1] https://docs.python.org/2/library/xml.etree.elementtree.html